requirements.txt
asgiref==3.4.1
Django==3.2.6
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Markdown==3.3.4
Pillow==8.3.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
